+++ test-server/CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## [0.2.4] - 2019-07-18
+
+* Update actix-server to 0.6
+
 ## [0.2.3] - 2019-07-16
 
 * Add `delete`, `options`, `patch` methods to `TestServerRunner`